requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appnope==0.1.2
    # via ipython
backcall==0.2.0
    # via ipython
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
decorator==4.4.2
    # via ipython
deprecated==1.2.11
    # via pygithub
gitdb==4.0.5
    # via gitpython
gitpython==3.1.12
    # via -r requirements.in
idna==2.10
    # via requests
ipython-genutils==0.2.0
    # via traitlets
ipython==7.20.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
parso==0.8.1
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.14
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pygithub==1.54.1
    # via -r requirements.in
pygments==2.7.4
    # via ipython
pyjwt==1.7.1
    # via pygithub
python-dotenv==0.15.0
    # via -r requirements.in
requests==2.25.1
    # via pygithub
smmap==3.0.5
    # via gitdb
traitlets==5.0.5
    # via ipython
urllib3==1.26.3
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.12.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
